requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.0
Flask-Migrate==4.0.4
PyMySQL==1.1.0
redis==5.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
